requirements.txt
backports.csv==1.0.7
beautifulsoup4==4.8.1
blis==0.4.1
certifi==2019.9.11
chardet==3.0.4
cheroot==8.2.1
CherryPy==18.3.0
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
en-core-web-lg==2.2.5
en-core-web-sm==2.2.5
et-xmlfile==1.0.1
feedparser==5.2.1
future==0.18.1
GetOldTweets3==0.0.11
graphviz==0.8.4
idna==2.6
importlib-metadata==0.23
jaraco.functools==2.0
jdcal==1.4.1
joblib==0.14.0
kiwisolver==1.1.0
lxml==4.4.1
matplotlib==3.1.1
more-itertools==7.2.0
murmurhash==1.0.2
mxnet==1.5.0
mysqlclient==1.4.4
nltk==3.4.5
numpy==1.16.6
oauthlib==3.1.0
openpyxl==3.0.3
pandas==0.25.2
Pattern==3.6
pdfminer.six==20191020
plac==1.1.3
portend==2.5
preshed==3.0.2
psutil==5.6.7
pycorenlp==0.3.0
pycryptodome==3.9.0
pyparsing==2.4.5
pyquery==1.4.1
PySocks==1.7.1
pyspellchecker==0.5.2
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.3
pywin32==225
requests==2.18.4
requests-oauthlib==1.2.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
soupsieve==1.9.4
spacy==2.2.2
srsly==0.2.0
stanfordcorenlp==3.9.1.1
tempora==1.14.1
thinc==7.3.1
tqdm==4.38.0
tweepy==3.8.0
urllib3==1.22
wasabi==0.4.0
zc.lockfile==2.0
zipp==0.6.0
